Use the unzip abstraction for partition
